избавился от слайса Favorites
