--- src/store/favorites/favorites.test.ts
@@ -1,37 +1,0 @@
-import { favorites } from './favorites';
-
-describe('Favorites Slice', () => {
-  it('should return initial state with empty action', () => {
-    const emptyAction = {type: ''};
-    const expectedState = {
-      favorites: [
-        {
-          'id': '6af6f711-c28d-4121-82cd-e0b462a27f00',
-          'title': 'Beautiful & luxurious studio at great location',
-          'type': 'apartment',
-          'price': 120,
-          'city': {
-            'name': 'Amsterdam',
-            'location': {
-              'latitude': 52.35514938496378,
-              'longitude': 4.673877537499948,
-              'zoom': 8
-            }},
-          'location': {
-            'latitude': 52.35514938496378,
-            'longitude': 4.673877537499948,
-            'zoom': 8
-          },
-          'isFavorite': false,
-          'isPremium': false,
-          'rating': 4,
-          'previewImage': 'https://url-to-image/image.png'
-        }
-      ]
-    };
-
-    const result = favorites.reducer(expectedState, emptyAction);
-
-    expect(result).toEqual(expectedState);
-  });
-});
