--- src/store/favorites/selectors.ts
@@ -1,5 +1,0 @@
-import { State } from '../../types/state';
-import { Offers } from '../../types/types';
-import { NameSpace } from '../const';
-
-export const getFavorites = (state: State): Offers => state[NameSpace.Favorites].favorites;
