+++ src/store/offers/offers.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { fetchNearOffersAction, fetchOffersAction, getOfferAction } from './thunks.ts';
+import { fetchFavoritesAction, fetchNearOffersAction, fetchOffersAction, getOfferAction } from './thunks.ts';
 import { NameSpace } from '../const';
 import { OfferInfo, Offers } from '../../types/types';
 
@@ -10,6 +10,7 @@ export type InitialState = {
   nearOffers: Offers;
   isOffersDataLoading: boolean;
   hasError: boolean;
+  favorites: Offers;
 };
 
 const initialState: InitialState = {
@@ -18,6 +19,7 @@ const initialState: InitialState = {
   nearOffers: [],
   isOffersDataLoading: false,
   hasError: false,
+  favorites: [],
 };
 
 export const offers = createSlice({
@@ -34,6 +36,20 @@ export const offers = createSlice({
         }
       });
     },
+    deleteFavorite: (state, action: PayloadAction<OfferInfo>) => {
+      for (let i = 0, len = state.favorites.length; i < len; i++) {
+        if (state.favorites[i].id === action.payload.id) {
+          state.favorites.splice(i, 1);
+          break;
+        }
+      }
+    },
+    addFavorite: (state, action: PayloadAction<OfferInfo>) => {
+      const favoriteOffer = state.offers.find((offer) => offer.id === action.payload.id);
+      if (favoriteOffer) {
+        state.favorites.push(favoriteOffer);
+      }
+    }
   },
   extraReducers(builder) {
     builder
@@ -60,8 +76,11 @@ export const offers = createSlice({
       })
       .addCase(fetchNearOffersAction.fulfilled, (state, action) => {
         state.nearOffers = action.payload;
+      })
+      .addCase(fetchFavoritesAction.fulfilled, (state, action) => {
+        state.favorites = action.payload;
       });
   },
 });
 
-export const {setOfferInfo, setFavoriteOffer} = offers.actions;
+export const {setOfferInfo, setFavoriteOffer, addFavorite, deleteFavorite} = offers.actions;
